requirements.txt
aiohttp
discord.py
psycopg2
PyYAML